backend/app/ai/requirements.txt
openai>=1.43.0
pdfplumber
PyPDF2
pillow
pytesseract
python-docx
